// src/frontend/src/components/operations/rollout/workplace-dialog/types.ts
export interface WorkplaceFormState {
  userName: string;
  userEmail: string;
  /** Entra ID (Azure AD Object ID) of the user */
  userEntraId: string;
  location: string;
  serviceId: number | undefined;
  scheduledDate: string | undefined;
  workplaceStatus: RolloutWorkplaceStatus;
  /** Linked physical workplace ID */
  physicalWorkplaceId: number | undefined;
  oldDevices: OldDeviceConfig[];
  configItems: AssetConfigItem[];
  returningOldDevice: boolean;
}